backend/requirements.txt
fastapi==0.116.1
greenlet==3.2.3
psycopg[binary]==3.2.9
pydantic==2.11.7
python-dotenv==1.1.1
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
